changed some other things to redice compiler warnings

+++ COSC363A1/base_object.cpp
@@ -6,7 +6,7 @@ void SceneObject::Render()
     
 }
 
-void SceneObject::Process(double dt)
+void SceneObject::Process(float dt)
 {
 
 }
